chapter6/lesson18/requirements.txt
# Lesson 18: AutoGen + MemGPT 智能体示例
# 多智能体协作和记忆管理

# 核心依赖
pyautogen>=0.2.0  # AutoGen 框架
memgpt>=0.3.0     # MemGPT 记忆管理

# LLM APIs
openai>=1.12.0
dashscope>=1.17.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# 工具和实用程序
python-dotenv>=1.0.0
tenacity>=8.2.0
requests>=2.31.0

# 可选依赖（用于高级功能）
# chromadb>=0.5.0  # 向量数据库
# redis>=5.0.0     # 缓存
# networkx>=3.1    # 图结构 